+++ packages/components/src/Label.js
@@ -1,16 +1,18 @@
 import React from 'react'
 import Box from './Box'
 
-export const Label = React.forwardRef((props, ref) => (
-  <Box
-    ref={ref}
-    as="label"
-    variant="label"
-    {...props}
-    __themeKey="forms"
-    __css={{
-      width: '100%',
-      display: 'flex',
-    }}
-  />
-))
+export const Label = React.forwardRef(function Label(props, ref) {
+  return (
+    <Box
+      ref={ref}
+      as="label"
+      variant="label"
+      {...props}
+      __themeKey="forms"
+      __css={{
+        width: '100%',
+        display: 'flex',
+      }}
+    />
+  )
+})